test: TST-035/037/046 + QNT-007 — stats functions, smoke_test, fp16 fix, INT8 outlier
TST-035: _stable_random_seed, _cohens_dz, _relative_gain_table,
         _build_paired_metric_rows, _paired_signflip_pvalue tests
TST-037: test_smoke_test.py (get_git_commit, get_hardware_info,
         CUDA-unavailable exit path)
TST-046: Fix fp16→fp32 dequant inconsistency in long-context GQA test
QNT-007: INT8 outlier quantization tests (clipping behavior)

+++ tests/test_triton_kernel.py
@@ -229,10 +229,12 @@ class TestTritonDecodeAttn(unittest.TestCase):
 
         k_reshaped = curr_k_int8.view(kv_heads, seq_len, num_groups, group_size)
         v_reshaped = curr_v_int8.view(kv_heads, seq_len, num_groups, group_size)
-        k_dequant_kv = (k_reshaped.to(torch.float16) * curr_k_scale.unsqueeze(-1)).view(
+        # TST-046: Use fp32 for dequant reference to match _torch_ref_decode (L61-62)
+        # and avoid fp16 precision loss on long contexts.
+        k_dequant_kv = (k_reshaped.to(torch.float32) * curr_k_scale.to(torch.float32).unsqueeze(-1)).view(
             kv_heads, seq_len, head_dim
         )
-        v_dequant_kv = (v_reshaped.to(torch.float16) * curr_v_scale.unsqueeze(-1)).view(
+        v_dequant_kv = (v_reshaped.to(torch.float32) * curr_v_scale.to(torch.float32).unsqueeze(-1)).view(
             kv_heads, seq_len, head_dim
         )
